renamed file to match the commands's name
